Reverse arguments of Node.compareDocumentPosition
When other library code may have wrapped nodes, such as to polyfill
Shadow DOM support, the reference node of the NodeIterator may
reject the argument to the compareDocumentPosition method with a
TypeError. By flipping the argument order, the wrapper has a chance
to polyfill this method to do the right thing.