Ensure error is thrown for empty text
When the traversal ends, ensure the reference node is always a text
node. If it is not, then no traversal was done in any direction, and no
normalization of the pointer was done between the forward and backward
traversals. For this to happen, the iterator must not contain any text
nodes in its collection.

This bug had been masked by a bug in dom-node-iterator@<4, wherein the
pointerBeforeReferenceNode property of the iterator was updated even
when no traversal occurred.